% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{createLink}
\alias{createLink}
\title{Create Hyperlink}
\usage{
createLink(author, link, title)
}
\arguments{
\item{author}{username of blog author}

\item{link}{link to create}

\item{title}{title visible name for link}
}
\value{
Data Frame with Details of Blog Posts
}
\description{
add a hyperlink to the data table
}
\examples{
None

}
