% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{flatten}
\alias{flatten}
\title{Flatten List}
\usage{
flatten(lst)
}
\arguments{
\item{lst}{list to flatten}
}
\value{
flattened list
}
\description{
utility function to flatten a multi level list
}
\examples{
None

}
