% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{formatNum}
\alias{formatNum}
\title{Format Number}
\usage{
formatNum(x)
}
\arguments{
\item{x}{The numeric character}
}
\value{
character representation of number
}
\description{
Format A Number as a Character with comma seperators
}
\examples{
None

}
