% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getAccount}
\alias{getAccount}
\title{getAccount}
\usage{
getAccount(username)
}
\arguments{
\item{username}{}
}
\value{
Account Details
}
\description{
Get the details of a user account
}
\examples{
getAccount()

}
