% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getAccountVotes}
\alias{getAccountVotes}
\title{getAccountVotes}
\usage{
getAccountVotes(user = "eroche")
}
\arguments{
\item{user}{Voting Account}
}
\value{
Data Table with list of votes and what they voted on
}
\description{
getAccountVotes
}
\examples{
getAccountVotes("eroche")

}
