% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getBlog}
\alias{getBlog}
\title{getBlog}
\usage{
getBlog(username)
}
\arguments{
\item{username}{Username of Blog Author.}
}
\value{
Data.Table with Details of Blog Posts
}
\description{
Get all main posts from a users blog (excluding Resteems)
}
\examples{
getBlog("eroche")

}
