% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getBlogPosts}
\alias{getBlogPosts}
\title{Blog Posts}
\usage{
getBlogPosts(username, node)
}
\arguments{
\item{username}{Username of Blog Author.}

\item{node}{Optional Argument (Hive Node to Query)}
}
\value{
Data.Table with Details of Blog Posts
}
\description{
Get all main posts from a users blog (excluding reHives). This function may take some time to process depending on the size of the blog history on an account
}
\examples{
getBlog("eroche")

}
