% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getComments}
\alias{getComments}
\title{Comment List}
\usage{
getComments(username, node)
}
\arguments{
\item{username}{Username of Comment Author.}

\item{node}{Optional Argument (Steem Node to Query)}
}
\value{
Data.Table with Details of Comments Posts
}
\description{
Get a list of all the comments from a users blog. This function may take some time to process depending on the size of the transaction history on an account
}
\examples{
getComments("eroche")

}
