% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getDelegation}
\alias{getDelegation}
\title{getDelegation}
\usage{
getDelegation(user = "eroche")
}
\arguments{
\item{user}{Account To Query}
}
\value{
Data Table with list of delegations
}
\description{
getDelegation
}
\examples{
getDelegation("eroche")

}
