% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getDelegations}
\alias{getDelegations}
\title{HP Delegations Made by an User}
\usage{
getDelegations(user = "eroche", node)
}
\arguments{
\item{user}{Account To Query}

\item{node}{Optional Argument (Hive Node to Query)}
}
\value{
Data Table with list of delegations
}
\description{
HP Delegations Made by an User
}
\examples{
getDelegation("eroche")

}
