% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getHiveProperties}
\alias{getHiveProperties}
\title{Hive Proporties}
\usage{
getHiveProperties(node)
}
\arguments{
\item{node}{Optional Argument (Hive Node to Query)}

\item{NULL}{}
}
\value{
Hive properties
}
\description{
Get details about the latest state of the Hive Blockchain.
}
\examples{
getHiveProperties()

}
