% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getPost}
\alias{getPost}
\title{getPost}
\usage{
getPost(username, permlink)
}
\arguments{
\item{username}{Username of Blog Author.}

\item{permlink}{Permlink of Blog Post}
}
\value{
List with Details of Post
}
\description{
Get details of a post
}
\examples{
getPost("eroche")

}
