% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getPosts}
\alias{getPosts}
\title{Get Steem Posts}
\usage{
getPosts(query)
}
\arguments{
\item{query}{get_discussions_by_author_before_date query}
}
\value{
list of blog posts
}
\description{
utility function to get the Steem posts from a user profile. This function is a wrapper to an RPC call.
}
\examples{
None

}
