% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getReplies}
\alias{getReplies}
\title{Replies Made to a Post}
\usage{
getReplies(user, permlink, node)
}
\arguments{
\item{user}{Account To Query}

\item{permlink}{Permlink of post to query}

\item{node}{Optional Argument (Steem Node to Query)}
}
\value{
Data Table with list of replies made to specified post
}
\description{
Replies Made to a Post
}
\examples{
getReplies("eroche", "data-wrangling-with-r")

}
