% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getSteemProperties}
\alias{getSteemProperties}
\title{getSteemProperties()}
\usage{
getSteemProperties()
}
\arguments{
\item{NULL}{}
}
\description{
Get the count of the accounts on the Steem Blockchain
}
\examples{
getSteemProperties()

}
