% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getTransactions}
\alias{getTransactions}
\title{getTransactions}
\usage{
getTransactions(user = "eroche", n = 100)
}
\arguments{
\item{user}{Account To Query}
}
\value{
The Last n number of transactions on an account
}
\description{
getTransactions
}
\examples{
getTransactions("eroche", 100)

}
