% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getWitnesses}
\alias{getWitnesses}
\title{getWitnesses}
\usage{
getWitnesses(limit = 1000)
}
\arguments{
\item{limit}{Number of Witnesses to Return

Get details of the witnesses}
}
\value{
List with Details of Witnesses
}
\description{
getWitnesses
}
\examples{
getWitnesses() Get first 1000 Witnesses
getWitnesses(20) Get Top 20 Witnesses

}
