% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{loopPosts}
\alias{loopPosts}
\title{Loop Through Blog}
\usage{
loopPosts(username)
}
\arguments{
\item{username}{Username of Blog Author}
}
\value{
Data Frame with Details of Blog Posts
}
\description{
Loop throught all the posts of a users blog
}
\examples{
loopPosts()

}
