% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{reputation.formatter}
\alias{reputation.formatter}
\title{Steem Reputation Formatter}
\usage{
reputation.formatter(rep)
}
\arguments{
\item{rep}{long form value}
}
\value{
2 digit value
}
\description{
utility convert the Steem Reputation to a 2 digit number. Ported Code to R to format rep in readable format https://github.com/steemit/steem-js/blob/master/src/formatter.js
}
\examples{
None

}
