% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getAccount}
\alias{getAccount}
\title{USer Account Details}
\usage{
getAccount(username, node)
}
\arguments{
\item{username}{}

\item{node}{Optional Argument (Hive Node to Query)}
}
\value{
Key Elements of Account Details; created, rep, post count, hive balance, savings balance, vesting
}
\description{
Get the details of a user account
}
\examples{
getAccount("eroche")

}
