% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getAccount}
\alias{getAccount}
\title{Account Details}
\usage{
getAccount(username, node)
}
\arguments{
\item{username}{}

\item{node}{Optional Argument (Steem Node to Query)}
}
\value{
Account Details
}
\description{
Get the details of a user account
}
\examples{
getAccount("eroche")

}
