% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getAccountVotes}
\alias{getAccountVotes}
\title{Votes made by an Account}
\usage{
getAccountVotes(user = "eroche", node)
}
\arguments{
\item{user}{Account to query}

\item{node}{Optional Argument (Steem Node to Query)}
}
\value{
Data Table with list of votes and what they voted on
}
\description{
Get a list of all the votes made by an account
}
\examples{
getAccountVotes("eroche")
getAccountVotes("eroche", node="https://api.steemit.com")

}
