% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getDelegation}
\alias{getDelegation}
\title{Delegations Made by an Account}
\usage{
getDelegation(user = "eroche", node)
}
\arguments{
\item{user}{Account To Query}

\item{node}{Optional Argument (Steem Node to Query)}
}
\value{
Data Table with list of delegations
}
\description{
Delegations Made by an Account
}
\examples{
getDelegation("eroche")

}
