% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getNodes}
\alias{getNodes}
\title{List of Steem RPC nodes}
\usage{
getNodes()
}
\arguments{
\item{NULL}{}
}
\value{
List of nodes
}
\description{
The functions in steemRdata use public RPC nodes. This function returns a list of possible nodes which can be specified in the function calls.
}
\examples{
getNodes()
}
