% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getPost}
\alias{getPost}
\title{Details of a Steem Post}
\usage{
getPost(username, permlink, node)
}
\arguments{
\item{username}{Username of Post Author.}

\item{permlink}{Permlink of Post}

\item{node}{Optional Argument (Steem Node to Query)}
}
\value{
List with Details of Post
}
\description{
Get details of a post, specified by the unique link (username and permlink)
}
\examples{
getPost("eroche", "time-series-with-r")
getPost("eroche", "time-series-with-r", "https://api.steemit.com")

}
